requirements.txt
pandas==2.3.3
numpy==2.2.6
pillow==12.0.0
tqdm==4.67.1
torch==2.9.1
torchvision==0.24.1
sentence-transformers==5.2.0
faiss-cpu==1.13.1
opencv-python==4.12.0.88
python-telegram-bot==22.5
transformers==4.57.3
timm==1.0.22
accelerate==1.12.0
aiogram==3.23.0
langchain==1.2.0
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
python-dotenv==1.2.1

